add: single entry point to mainWindow
unsafe to use

v0.0.4
