--- .idea/.copilot-plugin
@@ -1,6 +1,0 @@
-#Vaadin Copilot Integration Runtime Properties
-#Fri Nov 21 19:06:43 IST 2025
-endpoint=http\://127.0.0.1\:63342/api/copilot-c56315a2-9536-4ed4-888f-44a79aaece39
-ide=intellij
-supportedActions=write,writeBase64,delete,undo,redo,refresh,showInIde,getModulePaths,compileFiles,restartApplication,getVaadinRoutes,getVaadinVersion,getVaadinComponents,getVaadinEntities,getVaadinSecurity,reloadMavenModule,heartbeat
-version=1.5.5
